Added first working version of pydqed module.
The pydqed module is a Cython module that acts as the interface between
Python and DQED. This isn't the only way to write such a wrapper, but
Cython is particularly nice because the end user can later accelerate
their function evalutions (which are probably slow in pure Python) using
Cython.

The key method of interaction for the end user is the DQED class. Users
derive from this class and implement the evaluate method to provide the
residual and Jacobian values. Extra values needed by the evaluate method
can be stored as attributes of the derived class. The pydqed module
automatically wraps the evaluate method so that it is called when needed
by DQED.