Added fnorm and fjac attributes and dqed() method to DQED class.
Adding fnorm and fjac as attributes means that the user will have access
to their final values after the call to DQED is complete. I've also
fixed bugs in the initialization of fjac (it must be stored in Fortran
order since it is a matrix, and the number of columns was off by 1).

The dqed() method simply encapsulates the ugly code that calls into DQED
so that the solve() method is more readable. Originally I was going to
use it to query DQED for the proper sizes of iwork and rwork, but this
caused an error message to be printed each time, which the end user
will think is an error.