Now properly handing the case where linear constraints are used.
Lower and upper bounds must be specified for the constraint, and the
residual and Jacobian of that constraint must be evaluated and placed in
the proper location in the fj array. Note that, since we've made fjac an
attribute of the DQED class, we can access the memory that fj uses via
fjac, which makes things a bit more readable.