Some changes to how the length of rwork and iwork are determined.
I don't think the formulas given in the DQED comments are quite correct,
although they appear to be close. After stepping through the actual code
I've changed a few bits to hopefully be closer to correct. I'm not sure
I quite have it, though, especially when one or more constraints is
used.